Remove ContributionTracking Tests and Update DB
Testing on ContributionTracking is apparently so broken as to
not be worth it. We're rolling this into DI eventually anyways
so no huge loss.

However -- in the meantime for testing friendlyness elsewhere...

Don't always require an external DB.

And note that the patch files, whilst not SQLite safe, are not needed
because the ContributionTracking.sql file is up to date.

Change-Id: I0e583b33e3c9413cb04b95c4c238b3ddbd843764